+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Added `craft\base\Element::EVENT_PREP_QUERY_FOR_TABLE_ATTRIBUTE`. ([#9862](https://github.com/craftcms/cms/pull/9862))
 - Added `craft\base\ElementInterface::getIsFresh()`.
 - Added `craft\base\ElementInterface::setIsFresh()`.
-- Added `craft\events\PrepareElementQueryForTableAttributeEvent`.
+- Added `craft\events\ElementIndexTableAttributeEvent`.
 
 ### Changed
 - Enhanced the visibility of global navigation badges.